修改用例

+++ src/main/java/com/haisheng/framework/testng/bigScreen/jiaochen/wm/sense/pc/receptionmanage/ReceptionPurchaseFixedPackageScene.java
@@ -9,7 +9,7 @@ import lombok.Builder;
  */
 @Builder
 public class ReceptionPurchaseFixedPackageScene extends BaseScene {
-    private final Integer receptionId;
+    private final Long receptionId;
     private final Long customerId;
     private final String customerPhone;
     private final Long packageId;
